Create README.md
Need help with class assignment. My readFile method is not assigning values to the elements of the array from the txt file. Or, not how it should any way. I do not understand. The getLines method returns the correct amount of lines. But, the toString also returns the wrong data from the file. Both the readFile and toString seem to only reference the last number in the file but, give a zero value for the rest. Thank you for any feedback.